requirements.txt
pinecone
openai
python-dotenv
pyTelegramBotAPI
haystack-ai
requests
docstring-parser
loguru